+++ src/MetopDatasets.jl
@@ -17,6 +17,7 @@ include("genericFunctions/generic_functions.jl")
 include("auto_generate_tools/auto_generate_tool.jl")
 include("MetopDiskArray/MetopDiskArray.jl")
 include("MetopDiskArray/FlexibleMetopDiskArray.jl")
+include("MetopDiskArray/construct_disk_array.jl")
 include("InterfaceDataModel/InterfaceDataModel.jl")
 
 # Instruments